Update Type Match / Distance System
- Removed redundant and non-transitive `getSub…` methods from the
`IType` interface.
- Added the `Types.isAssignable(IType, IType)`, `.getDistance(…)` and
`.isConvertible(…)` methods.
- Changed the `IValue.getTypeMatch` return type from `float` to `int`.
- Renamed the `IType.classEquals(IType)` method to `isSameClass`.

+++ src/compiler/dyvil/tools/compiler/ast/access/FieldAssignment.java
@@ -137,7 +137,7 @@ public final class FieldAssignment implements IValue, INamed, IReceiverAccess, I
 	}
 	
 	@Override
-	public float getTypeMatch(IType type)
+	public int getTypeMatch(IType type)
 	{
 		return this.value.getTypeMatch(type);
 	}
