+++ src/compiler/dyvil/tools/compiler/ast/constant/WildcardValue.java
@@ -60,7 +60,7 @@ public class WildcardValue implements IConstantValue
 	}
 	
 	@Override
-	public float getTypeMatch(IType type)
+	public int getTypeMatch(IType type)
 	{
 		return 1;
 	}
