+++ src/compiler/dyvil/tools/compiler/ast/statement/StatementList.java
@@ -170,7 +170,7 @@ public class StatementList implements IValue, IValueList, IDefaultContext, ILabe
 	}
 
 	@Override
-	public float getTypeMatch(IType type)
+	public int getTypeMatch(IType type)
 	{
 		if (this.valueCount > 0)
 		{
